CRM-14931 - CRM.loadPage - Dialogs should respect statusBounce
When a dialog requests a page that sends a status-bounce, show the alert and
close the dialog.  I'm not sure if this policy is overly broad (I don't
know all the ways in which CRM.loadPage gets used), but it feels like
an improvement on the "Manage Case" screen.

It's still not ideal -- an empty dialog briefly flashes on the screen while
the AJAX request is sent.  But we do need *some* visual feedback about the
pending AJAX request, and that's currently what we have.