(Proof of Concept) CRM_Core_BAO_Cache - Encrypt persistent form data
The form layer (temporarily) stores data in the `civicrm_cache` (by way of a
session variable).  If the form contains sensitive data, then a
vulnerability that exposes cache data (e.g.  SQL injection or misplaced
backup) could compromise any sensitive information from the forms.

With this patch, the form data is still stored (temporarily), but it is
crypted with a two-part key.  To encrypt or decrypt, you must have both
parts of the key:

 * Site Key (`CIVICRM_SITE_KEY`) - This is private to the CiviCRM installation
 * Terminal Key (`cvtk`) - This is private to the user, relayed via cookie.

For an attacker to reda the form data, he would need three things -- the raw
database, the site key, and each users' terminal key.

Open topics/issues:
 * How does the extra crypto affect performance?
 * What are the exact parameters for the cookie?
 * Stress-test edge-cases (e.g. what happens if the `cvtk` cookie expires before the session cookie?)
 * Some kind of test coverage
 * How well php-sec-lib (`Crypt_RSA`, etc) perform across our user-community? (If you have php-openssl
   installed, it should be pretty good. But if you don't have that, then it uses polyfilled implementations.)