CRM_Core_Error - Change/simplify debug_var, debug_log_message signatures
The signatures for these functions specify that the final log message should
be returned as a string. This would become ill-defined if you changed the output
format, and it semantically doesn't comply with PSR-3.

After grepping universe, I can't find anything that actually relies on these
functions returning the string. Which makes sense: as someone submitting a
log, I want to send a forget.

This simplifies the contract and makes it easier to refactor.