Allow sites to accept REST API calls without the site-key (Demo/Untested)
Arguably, it's redundant to require both a general "site-key" and a per-user
"api-key". However, demonstrating this is a tricky. (If we're wrong, or if
some site has a bad api-key policy, then we wouldn't want the entire ecosystem
to be put at risk.)

This approach allows folks to incrementally transition towards the simpler
policy based on their own cost/benefit.