MailingAB.submit - Accept `winner_id` option
When a "winner" is selected for an A/B test, several fields must be copied over.
This provides the "copy" logic on the server-side.

Before
-------

The copy logic is only available in Javascript, so it's hard to reference in a unit-test.

After
-----

The copy logic is available as an extra option for the MailingAB.submit API.
By default, the API works as before.  If you add the option `winner_id`,
then it will also copy-over the details from the winning experiment.