$config->{configAndLogDir,templateCompileDir} - Move from "Runtime" to "Paths"
This slightly changes the computation of the private log and compilation
directories so that:

* The definition appears more declarative.
* The default is computed on-demand (i.e. later in the process).
* The values can be overriden more intuitively.

Before
------

The properties `$config->configAndLogDir` and `$config->templateCompileDir`
are computed as part of a procedural initialization step in
`CRM_Core_Config_Runtime`.

After
-----

The properties `$config->configAndLogDir` and `$config->templateCompileDir`
are computed on-demand by `Civi\Core\Paths`.

The values of these properties may:

* Be computed automatically in a way that matches the previous output
  (i.e.  `configAndLogDir`=~`CIVICRM_TEMPLATE_COMPILEDIR/../configAndLog`;
  `templateCompileDir`=~`CIVICRM_TEMPLATE_COMPILEDIR`).
* Be customized as a pair (by setting `civicrm.private`)
* Be customized individually (by setting `civicrm.logs` or `civicrm.compile`).