(REF) APIv4 - Extract CheckRequiredTrait
This is an incremental step towards moving the 'required fields check' to be
an event/subscriber (`civi.api4.validate`).

Before: `AbstractAction::checkRequiredFields()` is a helper method on all
action classes. However, it is only used by two classes.

After: `CheckRequiredTrait` is a helper that can be used to mix-in
`checkRequired()` for a couple classes.

Why: The aim is to move this to its own subscriber, but that's a bit more
work right now.  This isolates the functional so that it's less likely to
get accidental usage and so that it's easier to see the lines around it.