WorkflowMessage - Add base class, interface, test, field-spec
A `WorkflowMessage` is a data-model which describes the parameters required
to send an automated worklow message.

Formally, a `WorkflowMessage` is an interface and does not need to be modeled
as a class.  (Thus, `getFields()`, `import()`, `export()` work with arrays.)
However, the standard implementations are based on class models (based on
examining the properties/annotations).