AssetBuilder - Fix `testInvalid()` failure. Strengthen the digest code.
After updating `AssetBuilder` to validate the digest-code, we started to get
failures in `E2E\Core\AssetBuilderTest`.  Why?

The digest code relies on a (semi-)secret value, the runtime-ID.  This ID is
different in `phpunit8` and `php-fpm` processes.  (Specifically, I believe
that `SCRIPT_FILENAME` throws it off.) Thus, the PHPUnit test generates one
digest-code, but the HTTP server expects a different digest-code.

This function has a "TODO" to revise the digest code. Resolving this TODO
also resolves the test discrepancy.

There is just one catch: resolving the TODO requires a secret-key. We
have specified a suitable secret-key, but some upgraded sites may not
have deployed a secret-key. For this edge-case, we keep the old code
as a fallback.