Drop CIVICRM_INSTALLED in favor of pre-existing CIVICRM_SETTINGS_PATH
In the preceding commits of the PR, it updates `civicrm.settings.php` to add a new constant
CIVICRM_INSTALLED.  The constant is simply TRUE.  In other words,

> CIVICRM_INSTALLED is TRUE if-and-only-if civicrm.settings.php has been loaded

Of course, deploying new constants in `civicrm.settings.php` can be annoying.
Do we need to?  Well, there's already a similar constant...  sort of.

- Drupal, Backdrop, WordPress, and Joomla all set the constant CIVICRM_SETTINGS_PATH
  when they load `civicrm.settings.php`. You can check this variable to see
  if the settings are loaded.
- But Standalone doesn't currently. If we make it conform, then we don't
  need to add CIVICRM_INSTALLED.