(POC) component-bundle-json@1 - Render dynamic bundles of components
This sort of works, but sort of not.

WebComponents are registered via `customElements.define()`; they don't need
to be `export`ed.  So, in theory, those files can be concatenated.

The hope of this commit was to use asset-builder for dynamic/automatic
concatentation...  and then rely on the importmap to ensure that nested
`import`s work.

The validity depends on how you formulate the nested `import`. Compare:

- import greetings from "../greetings.js";
- import greetings from "org.civicrm.afform/js/greetings.js";

This form of bundling will break on the relative-path (`../`). (The browser
evaluates the relative-path against the physical URL of the requester.)

But it does work with symbolic-path (`org.civicrm.afform/`)...